+++ lib/index.js
@@ -10,6 +10,7 @@ const url = require('url')
 const yaml = require('js-yaml')
 
 const handler = require('./lambda/handler')
+const taskPlan = require('./lambda/taskPlan')
 const npm = require('./npm')
 const Serverless = require('./serverless-fx')
 
@@ -134,8 +135,8 @@ const impl = {
     const networkBuffer = 2 // seconds to allow for network transmission
     const resultsBuffer = 3 // seconds to allow for processing overhead (validation, decision making, results calculation)
     const settings = handler.impl.getSettings(script)
-    const durationInSeconds = handler.impl.scriptDurationInSeconds(script)
-    const requestsPerSecond = handler.impl.scriptRequestsPerSecond(script)
+    const durationInSeconds = taskPlan.impl.scriptDurationInSeconds(script)
+    const requestsPerSecond = taskPlan.impl.scriptRequestsPerSecond(script)
     let httpTimeout = 120 // default AWS setting
     if (
       aws.config &&
