+++ tests/lib/lambda/handler.spec.js
@@ -1,7 +1,5 @@
-const artillery = require('artillery-core')
 const aws = require('aws-sdk')
 const chai = require('chai')
-const EventEmitter = require('events')
 const quibble = require('quibble')
 const path = require('path')
 const sinon = require('sinon')
@@ -18,7 +16,6 @@ const lambdaInvokeStub = sinon.stub(aws.Service.prototype, 'makeRequest')
 const handler = require(path.join('..', '..', '..', 'lib', 'lambda', 'handler.js'))
 
 let script
-let phase
 let result
 let expected
 
@@ -26,7 +23,7 @@ const tagContext = {
   functionName: 'name',
 }
 
-const validScript = () => ({
+const tagScript = () => ({
   config: {
     phases: [
       { duration: 1, arrivalRate: 2 },
@@ -112,257 +109,17 @@ describe('./lib/lambda/handler.js', () => {
       })
     })
 
-    describe('#phaseDurationInSeconds', () => {
-      it('extracts the duration of a constant rate phase', () => {
-        phase = {
-          duration: 10,
-          arrivalRate: 10,
-        }
-        expect(handler.impl.phaseDurationInSeconds(phase)).to.eql(phase.duration)
-      })
-      it('extracts the duration of a ramp to phase', () => {
-        phase = {
-          duration: 10,
-          arrivalRate: 10,
-          rampTo: 20,
-        }
-        expect(handler.impl.phaseDurationInSeconds(phase)).to.eql(phase.duration)
-      })
-      it('extracts the duration of a pause phase', () => {
-        phase = {
-          pause: 10,
-        }
-        expect(handler.impl.phaseDurationInSeconds(phase)).to.eql(phase.pause)
-      })
-      it('extracts the duration of a non-phase', () => {
-        phase = {}
-        expect(handler.impl.phaseDurationInSeconds(phase)).to.eql(-1)
-      })
-    })
-
-    describe('#scriptDurationInSeconds', () => {
-      it('detects and reports the index of invalid phases in a script', () => {
-        script = {
-          config: {
-            phases: [
-              {}, // not valid
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(-0)
-      })
-      it('detects and reports invalid first phase in a script', () => {
-        script = {
-          config: {
-            phases: [
-              {}, // not valid
-              { duration: 10, arrivalRate: 10 },
-              { duration: 10, arrivalRate: 10 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(-0)
-      })
-      it('detects and reports the index of invalid phases in a script', () => {
-        script = {
-          config: {
-            phases: [
-              { duration: 10, arrivalRate: 10 },
-              {}, // not valid
-              { duration: 10, arrivalRate: 10 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(-1)
-      })
-      it('detects and reports an invalid last phase in a script', () => {
-        script = {
-          config: {
-            phases: [
-              { duration: 10, arrivalRate: 10 },
-              { duration: 10, arrivalRate: 10 },
-              {}, // not valid
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(-2)
-      })
-      it('detects and reports the first index of invalid phases in a script', () => {
-        script = {
-          config: {
-            phases: [
-              { duration: 10, arrivalRate: 10 },
-              { duration: 10, arrivalRate: 10 },
-              {}, // not valid
-              { duration: 10, arrivalRate: 10 },
-              { duration: 10, arrivalRate: 10 },
-              {}, // also not valid
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(-2)
-      })
-      it('correctly calculates the duration of a single-phase script', () => {
-        script = {
-          config: {
-            phases: [
-              { duration: 10 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(10)
-      })
-      it('correctly calculates the duration of a multi-phase script', () => {
-        script = {
-          config: {
-            phases: [
-              { duration: 1 },
-              { duration: 2 },
-              { duration: 3 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptDurationInSeconds(script)).to.eql(6) // 1 + 2 + 3 = 6
-      })
-      it('correctly calculates the duration of this specific script', () => {
-        script = {
-          config: {
-            phases: [
-              {
-                duration: 60,
-                arrivalRate: 1,
-                rampTo: 5,
-                name: 'Ramp-Up',
-              },
-              {
-                duration: 120,
-                arrivalRate: 5,
-                name: 'Full-Load',
-              },
-              {
-                duration: 60,
-                arrivalRate: 5,
-                rampTo: 1,
-                name: 'Ramp-Down',
-              },
-            ],
-          },
-        }
-        result = handler.impl.scriptDurationInSeconds(script)
-        expect(result).to.equal(240)
-      })
-    })
-
-    describe('#phaseRequestsPerSecond', () => {
-      it('calculates the arrivals per second specified in an ascending rampTo phase', () => {
-        phase = { arrivalRate: 10, rampTo: 20 }
-        expect(handler.impl.phaseRequestsPerSecond(phase)).to.eql(20)
-      })
-      it('calculates the arrivals per second specified in an descending rampTo phase', () => {
-        phase = { arrivalRate: 20, rampTo: 10 }
-        expect(handler.impl.phaseRequestsPerSecond(phase)).to.eql(20)
-      })
-      it('calculates the arrivals per second in a constant rate phase', () => {
-        phase = { arrivalRate: 20 }
-        expect(handler.impl.phaseRequestsPerSecond(phase)).to.eql(20)
-      })
-      it('calculates the arrivals per second in a constant count phase', () => {
-        phase = { arrivalCount: 20, duration: 1 }
-        expect(handler.impl.phaseRequestsPerSecond(phase)).to.eql(20)
-      })
-      it('calculates the arrivals per second in a pause phase', () => {
-        phase = { pause: 1 }
-        expect(handler.impl.phaseRequestsPerSecond(phase)).to.eql(0)
-      })
-      it('returns -1 to indicate an invalid phase', () => {
-        phase = {}
-        expect(handler.impl.phaseRequestsPerSecond(phase)).to.eql(-1)
-      })
-    })
-
-    describe('#scriptRequestsPerSecond', () => {
-      it('correctly identifies an invalid phase in a script', () => {
-        script = {
-          config: {
-            phases: [
-              {}, // not valid
-            ],
-          },
-        }
-        expect(handler.impl.scriptRequestsPerSecond(script)).to.eql(-0)
-      })
-      it('correctly calculates the index of an invalid phase in a script when it is the first phase', () => {
-        script = {
-          config: {
-            phases: [
-              {}, // not valid
-              { arrivalRate: 1 },
-              { arrivalRate: 2 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptRequestsPerSecond(script)).to.eql(-0)
-      })
-      it('correctly calculates the index of an invalid phase in a script when it the phase is in the midst of the phases', () => {
-        script = {
-          config: {
-            phases: [
-              { arrivalRate: 1 },
-              {}, // not valid
-              { arrivalRate: 2 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptRequestsPerSecond(script)).to.eql(-1)
-      })
-      it('correctly calculates the index of an invalid phase in a script when it is the last phase', () => {
-        script = {
-          config: {
-            phases: [
-              { arrivalRate: 1 },
-              { arrivalRate: 2 },
-              {}, // not valid
-            ],
-          },
-        }
-        expect(handler.impl.scriptRequestsPerSecond(script)).to.eql(-2)
-      })
-      it('correctly calculates the arrivals per second of a single-phase script', () => {
-        script = {
-          config: {
-            phases: [
-              { arrivalRate: 1 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptRequestsPerSecond(script)).to.eql(1)
-      })
-      it('correctly calculates the arrivals per second of a multi-phase script', () => {
-        script = {
-          config: {
-            phases: [
-              { arrivalRate: 1 },
-              { arrivalRate: 2 },
-              { arrivalRate: 3 },
-              { arrivalRate: 4 },
-              { arrivalRate: 5 },
-            ],
-          },
-        }
-        expect(handler.impl.scriptRequestsPerSecond(script)).to.eql(5)
-      })
-    })
     describe('#validScript', () => {
       /* eslint-disable no-underscore-dangle */
       beforeEach(() => {
-        script = validScript()
+        script = tagScript()
       })
       it('accepts the valid script', () => {
         expect(handler.impl.validScript(script, null, noop)).to.be.true
       })
       describe('_split usage', () => {
         const validSplitScript = () => {
-          const ret = validScript()
+          const ret = tagScript()
           ret._split = {}
           return ret
         }
@@ -462,652 +219,7 @@ describe('./lib/lambda/handler.js', () => {
         })
       })
     })
-    /**
-     * SPLIT PHASE BY DurationInSeconds
-     */
-    describe('#splitPhaseByDurationInSeconds splits PHASES that are TOO LONG', () => {
-      it('splitting a constant rate phase of three chunk\'s durationInSeconds into two parts.', () => {
-        phase = {
-          duration: 3 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          arrivalRate: 1,
-        }
-        expected = {
-          chunk: {
-            duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalRate: 1,
-          },
-          remainder: {
-            duration: 2 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalRate: 1,
-          },
-        }
-        result = handler.impl.splitPhaseByDurationInSeconds(phase, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting a ramping phase of two chunk\'s duration into two parts.', () => {
-        phase = {
-          duration: 3 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          arrivalRate: 1,
-          rampTo: 4,
-        }
-        expected = {
-          chunk: {
-            duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalRate: 1,
-            rampTo: 2,
-          },
-          remainder: {
-            duration: 2 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalRate: 2,
-            rampTo: 4,
-          },
-        }
-        result = handler.impl.splitPhaseByDurationInSeconds(phase, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splits ramp of two chunk\'s duration into two parts, rounding to the nearest integer rate per second.', () => {
-        phase = {
-          duration: 2 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          arrivalRate: 1,
-          rampTo: 2,
-        }
-        expected = {
-          chunk: {
-            duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalRate: 1,
-            rampTo: 2,
-          },
-          remainder: {
-            duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalRate: 2,
-            rampTo: 2,
-          },
-        }
-        result = handler.impl.splitPhaseByDurationInSeconds(phase, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting an arrival count phase of two chunk\'s duration into two parts.', () => {
-        phase = {
-          duration: 2 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          arrivalCount: 2 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-        }
-        expected = {
-          chunk: {
-            duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalCount: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          },
-          remainder: {
-            duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-            arrivalCount: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          },
-        }
-        result = handler.impl.splitPhaseByDurationInSeconds(phase, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting a pause phase of two chunk\'s duration into two parts.', () => {
-        phase = {
-          pause: 2 * handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-        }
-        expected = {
-          chunk: {
-            pause: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          },
-          remainder: {
-            pause: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-          },
-        }
-        result = handler.impl.splitPhaseByDurationInSeconds(phase, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-    })
-    /**
-     * SPLIT SCRIPT BY DURATION
-     */
-    describe('#splitScriptByDurationInSeconds The handler splits SCRIPTS that are TOO LONG', () => {
-      it('splitting a script where there is a natural phase split at MAX_CHUNK_DURATION between two phases.', () => {
-        script = {
-          config: {
-            phases: [
-              {
-                duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-                arrivalRate: 1,
-              },
-              {
-                duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-                arrivalRate: 2,
-              },
-            ],
-          },
-        }
-        expected = {
-          chunk: {
-            config: {
-              phases: [
-                {
-                  duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-                  arrivalRate: 1,
-                },
-              ],
-            },
-          },
-          remainder: {
-            config: {
-              phases: [
-                {
-                  duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-                  arrivalRate: 2,
-                },
-              ],
-            },
-          },
-        }
-        result = handler.impl.splitScriptByDurationInSeconds(script, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting a script where there is a natural phase split at MAX_CHUNK_DURATION between many phases.', () => {
-        script = {
-          config: {
-            phases: [
-              {
-                duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                arrivalRate: 1,
-              },
-              {
-                duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                arrivalRate: 1,
-              },
-              {
-                duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                arrivalRate: 1,
-              },
-              {
-                duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                arrivalRate: 1,
-              },
-              {
-                duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-                arrivalRate: 2,
-              },
-            ],
-          },
-        }
-        expected = {
-          chunk: {
-            config: {
-              phases: [
-                {
-                  duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                  arrivalRate: 1,
-                },
-                {
-                  duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                  arrivalRate: 1,
-                },
-              ],
-            },
-          },
-          remainder: {
-            config: {
-              phases: [
-                {
-                  duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                  arrivalRate: 1,
-                },
-                {
-                  duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                  arrivalRate: 1,
-                },
-                {
-                  duration: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-                  arrivalRate: 2,
-                },
-              ],
-            },
-          },
-        }
-        result = handler.impl.splitScriptByDurationInSeconds(script, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting a script where a phase must be split at MAX_CHUNK_DURATION.', () => {
-        script = {
-          config: {
-            phases: [
-              {
-                duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.75),
-                arrivalRate: 1,
-              },
-              {
-                duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.75),
-                arrivalRate: 1,
-              },
-              {
-                duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.75),
-                arrivalRate: 1,
-              },
-            ],
-          },
-        }
-        expected = {
-          chunk: {
-            config: {
-              phases: [
-                {
-                  duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.75),
-                  arrivalRate: 1,
-                },
-                {
-                  duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.25),
-                  arrivalRate: 1,
-                },
-              ],
-            },
-          },
-          remainder: {
-            config: {
-              phases: [
-                {
-                  duration: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.5),
-                  arrivalRate: 1,
-                },
-                {
-                  duration: Math.ceil(handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS * 0.75),
-                  arrivalRate: 1,
-                },
-              ],
-            },
-          },
-        }
-        result = handler.impl.splitScriptByDurationInSeconds(script, handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS)
-        expect(result).to.deep.equal(expected)
-      })
-    })
-    /**
-     * SPLIT PHASE BY RequestsPerSecond
-     */
-    describe('#splitPhaseByRequestsPerSecond The handler splits PHASES that are TOO WIDE (RPS > MAX_RPS)', () => {
-      // min >= chunkSize
-      it('splitting a ramp phase that at all times exceeds a rate of DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND into a' +
-        ' constant rate and remainder ramp phases.', () => {
-        phase = {
-          arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-          rampTo: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 3,
-          duration: 1,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              rampTo: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-              duration: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      // max <= chunkSize
-      it('splitting ramp that at all times is less than DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND into a chunk of the ramp' +
-        ' phase and a remainder of a pause phase.', () => {
-        phase = {
-          arrivalRate: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.5),
-          rampTo: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.75),
-          duration: 1,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalRate: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.5),
-              rampTo: Math.floor(handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.75),
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              pause: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting an ascending ramp phase that starts lower than and ends higher than' +
-        ' DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND into a chunk of a ramp phase followed by a constant rate phase and a' +
-        ' remainder of a pause phase followed by a ramp phase.', () => {
-        phase = {
-          arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.5,
-          rampTo: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 1.5,
-          duration: 2,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalRate: phase.arrivalRate,
-              rampTo: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              pause: 1,
-            },
-            {
-              arrivalRate: 1,
-              rampTo: (handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 1.5) - handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND, // eslint-disable-line max-len
-              duration: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting a descending ramp phase that starts lower than and ends higher than' +
-        ' DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND into a chunk of a constant rate phase followed by a ramp phase and a' +
-        ' remainder of a ramp phase followed by a pause phase.', () => {
-        phase = {
-          arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 1.5,
-          rampTo: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.5,
-          duration: 2,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              rampTo: phase.rampTo,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              arrivalRate: phase.arrivalRate - handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              rampTo: 1,
-              duration: 1,
-            },
-            {
-              pause: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splits an arrivalRate of less than a chunk\'s requestsPerSecond into a chunk of that requestsPerSecond and remainder of pause', () => {
-        phase = {
-          arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.75,
-          duration: 1,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalRate: phase.arrivalRate,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              pause: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting an arrivalRate phase of two chunk\'s requestsPerSecond into two parts.', () => {
-        phase = {
-          arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-          duration: 1,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting an arrivalCount phase of two chunk\'s requestsPerSecond into two parts.', () => {
-        phase = {
-          arrivalCount: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-          duration: 1,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalCount: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              arrivalCount: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-              duration: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting an arrivalCount phase of less than chunkSize\'s requestsPerSecond into an arrival count and pause phase.', () => {
-        phase = {
-          arrivalCount: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 0.75,
-          duration: 1,
-        }
-        expected = {
-          chunk: [
-            {
-              arrivalCount: phase.arrivalCount,
-              duration: 1,
-            },
-          ],
-          remainder: [
-            {
-              pause: 1,
-            },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splitting a pause phase into two pause phases.', () => {
-        phase = {
-          pause: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS,
-        }
-        expected = {
-          chunk: [
-            { pause: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS },
-          ],
-          remainder: [
-            { pause: handler.constants.DEFAULT_MAX_CHUNK_DURATION_IN_SECONDS },
-          ],
-        }
-        result = handler.impl.splitPhaseByRequestsPerSecond(phase, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-    })
-    /**
-     * SPLIT SCRIPT BY RequestsPerSecond
-     */
-    describe('#splitScriptByRequestsPerSecond the handler splits SCRIPTS that are TOO WIDE', () => {
-      it('splits a script with a phase that specifies twice DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND load.', () => {
-        script = {
-          config: {
-            phases: [
-              {
-                duration: 1,
-                arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-              },
-            ],
-          },
-        }
-        expected = {
-          chunk: {
-            config: {
-              phases: [
-                {
-                  duration: 1,
-                  arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-                },
-              ],
-            },
-          },
-          remainder: {
-            config: {
-              phases: [
-                {
-                  duration: 1,
-                  arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-                },
-              ],
-            },
-          },
-        }
-        result = handler.impl.splitScriptByRequestsPerSecond(script, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-      it('splits a script of two phases that specify twice DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND load to split.', () => {
-        script = {
-          config: {
-            phases: [
-              {
-                duration: 1,
-                arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-              },
-              {
-                duration: 1,
-                arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND * 2,
-              },
-            ],
-          },
-        }
-        expected = {
-          chunk: {
-            config: {
-              phases: [
-                {
-                  duration: 1,
-                  arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-                },
-                {
-                  duration: 1,
-                  arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-                },
-              ],
-            },
-          },
-          remainder: {
-            config: {
-              phases: [
-                {
-                  duration: 1,
-                  arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-                },
-                {
-                  duration: 1,
-                  arrivalRate: handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND,
-                },
-              ],
-            },
-          },
-        }
-        result = handler.impl.splitScriptByRequestsPerSecond(script, handler.constants.DEFAULT_MAX_CHUNK_REQUESTS_PER_SECOND)
-        expect(result).to.deep.equal(expected)
-      })
-    })
 
-    describe('#loadProcessor', () => {
-      it('loads custom processor code based on script configuration', () => {
-        const newScript = {
-          config: {
-            processor: `${__dirname}/customprocessor.js`,
-          },
-        }
-        handler.impl.loadProcessor(newScript)
-        expect(newScript.config.processor.testMethod()).to.equal('testValue')
-      })
-      it('does not attempt to reload a previously loaded processor', () => {
-        const newScript = {
-          config: {
-            processor: {
-              f: () => 'testValue',
-            },
-          },
-        }
-        handler.impl.loadProcessor(newScript)
-        expect(newScript.config.processor.f()).to.equal('testValue')
-      })
-    })
-    describe('#readPayload', () => {
-      it('reads a single payload file.', () => {
-        const newScript = {
-          config: {
-            payload: {
-              path: path.join(__dirname, 'example.0.csv'),
-            },
-          },
-        }
-        const payload = handler.impl.readPayload(newScript)
-        expect(payload).to.deep.equal([
-          ['123456', 'John Doe'],
-          ['234567', 'Jane Doe'],
-          ['345678', 'Baby Doe'],
-        ])
-      })
-      it('reads a set of payload files.', () => {
-        const newScript = {
-          config: {
-            payload: [
-              {
-                path: path.join(__dirname, 'example.0.csv'),
-              },
-              {
-                path: path.join(__dirname, 'example.1.csv'),
-              },
-            ],
-          },
-        }
-        const payload = handler.impl.readPayload(newScript)
-        expect(payload).to.deep.equal([
-          {
-            path: path.join(__dirname, 'example.0.csv'),
-            data: [
-              ['123456', 'John Doe'],
-              ['234567', 'Jane Doe'],
-              ['345678', 'Baby Doe'],
-            ],
-          },
-          {
-            path: path.join(__dirname, 'example.1.csv'),
-            data: [
-              ['123457', 'John Jones'],
-              ['234568', 'Jane Jones'],
-              ['345679', 'Baby Jones'],
-            ],
-          },
-        ])
-      })
-    })
     /**
      * SPLIT SCRIPT BY FLOW
      */
@@ -1335,104 +447,6 @@ describe('./lib/lambda/handler.js', () => {
       })
     })
 
-    describe('#runLoad', () => {
-      let loadProcessorStub
-      let readPayloadStub
-      let runner
-      let runnerStub
-      beforeEach(() => {
-        loadProcessorStub = sinon.stub(handler.impl, 'loadProcessor').returns()
-        readPayloadStub = sinon.stub(handler.impl, 'readPayload').returns()
-        runner = new EventEmitter()
-        runner.run = noop
-        runnerStub = sinon.stub(artillery, 'runner').returns(runner)
-      })
-      afterEach(() => {
-        loadProcessorStub.restore()
-        readPayloadStub.restore()
-        runnerStub.restore()
-      })
-      it('does nothing in simulation mode',
-        () => new Promise((resolve) => {
-          script = { _trace: true, _simulation: true }
-          handler.impl.runLoad(1, script, tagContext, (err, res) => {
-            expect(err).to.be.null
-            expect(res).to.eql({ Payload: '{ "errors": 0 }' })
-            resolve()
-          })
-        }).then(() => {
-          expect(loadProcessorStub).to.not.be.called
-          expect(readPayloadStub).to.not.be.called
-          expect(runnerStub).to.not.be.called
-        }) // eslint-disable-line comma-dangle
-      )
-      it('"loads" a processor, "reads" a payload, and "executes" the script before the callback is executed in standard mode',
-        () => new Promise((resolve) => {
-          script = {}
-          const opts = {}
-          const report = {}
-          handler.impl.runLoad(1, script, tagContext, (err, res) => {
-            expect(err).to.be.null
-            expect(res).to.equal(report)
-            resolve()
-          })
-          runner.emit('phaseStarted', opts)
-          runner.emit('phaseCompleted', opts)
-          runner.emit('done', report)
-        }).then(() => {
-          expect(loadProcessorStub).to.be.calledOnce
-          expect(loadProcessorStub).to.be.calledBefore(readPayloadStub)
-          expect(readPayloadStub).to.be.calledAfter(loadProcessorStub)
-          expect(readPayloadStub).to.be.calledOnce
-          expect(readPayloadStub).to.be.calledBefore(runnerStub)
-          expect(runnerStub).to.be.calledAfter(readPayloadStub)
-          expect(runnerStub).to.be.calledOnce
-        }) // eslint-disable-line comma-dangle
-      )
-      it('"loads" a processor, "reads" a payload, and "executes" the script before the callback is executed in trace mode',
-        () => new Promise((resolve) => {
-          script = { _trace: true }
-          const opts = {}
-          const report = {}
-          handler.impl.runLoad(1, script, tagContext, (err, res) => {
-            expect(err).to.be.null
-            expect(res).to.equal(report)
-            resolve()
-          })
-          runner.emit('phaseStarted', opts)
-          runner.emit('phaseCompleted', opts)
-          runner.emit('done', report)
-        }).then(() => {
-          expect(loadProcessorStub).to.be.calledOnce
-          expect(loadProcessorStub).to.be.calledBefore(readPayloadStub)
-          expect(readPayloadStub).to.be.calledAfter(loadProcessorStub)
-          expect(readPayloadStub).to.be.calledOnce
-          expect(readPayloadStub).to.be.calledBefore(runnerStub)
-          expect(runnerStub).to.be.calledAfter(readPayloadStub)
-          expect(runnerStub).to.be.calledOnce
-        }) // eslint-disable-line comma-dangle
-      )
-      it('handled exepected errors in the artillery runner or runner event invocations',
-        () => new Promise((resolve) => {
-          script = {}
-          runnerStub.throws()
-          handler.impl.runLoad(1, script, tagContext, (err, res) => {
-            expect(err).to.be.a('string')
-            expect(res).to.be.undefined
-            resolve()
-          })
-        }).then(() => {
-          expect(loadProcessorStub).to.be.calledOnce
-          expect(loadProcessorStub).to.be.calledBefore(readPayloadStub)
-          expect(readPayloadStub).to.be.calledAfter(loadProcessorStub)
-          expect(readPayloadStub).to.be.calledOnce
-          expect(readPayloadStub).to.be.calledBefore(runnerStub)
-          expect(runnerStub).to.be.calledAfter(readPayloadStub)
-          expect(runnerStub).to.be.calledOnce
-        }) // eslint-disable-line comma-dangle
-      )
-    })
-
     describe('#runPerformance', () => {
       // /**
       //  * Run an Artillery script.  Detect if it needs to be split and do so if it does.  Execute scripts not requiring
@@ -1624,7 +638,7 @@ describe('./lib/lambda/handler.js', () => {
       expect(runPerformanceStub).to.not.have.been.called
     })
     it('adds _genesis if not present', () => {
-      script = validScript()
+      script = tagScript()
       expect(script._genesis).to.be.undefined
       handler.api.run(script, tagContext, noop)
       expect(script._genesis).to.be.a('number')
@@ -1634,7 +648,7 @@ describe('./lib/lambda/handler.js', () => {
     })
     it('maintains a given _genesis', () => {
       const genesis = 12345
-      script = validScript()
+      script = tagScript()
       script._genesis = genesis
       expect(script._genesis).to.be.equal(genesis)
       handler.api.run(script, tagContext, noop)
@@ -1645,7 +659,7 @@ describe('./lib/lambda/handler.js', () => {
     })
     it('detects acceptance mode declared as "acc"', () => {
       const mode = 'acc'
-      script = validScript()
+      script = tagScript()
       script.mode = mode
       handler.api.run(script, tagContext, noop)
       expect(validScriptStub).to.have.been.calledOnce
@@ -1654,7 +668,7 @@ describe('./lib/lambda/handler.js', () => {
     })
     it('detects acceptance mode declared as "acceptance"', () => {
       const mode = 'acceptance'
-      script = validScript()
+      script = tagScript()
       script.mode = mode
       handler.api.run(script, tagContext, noop)
       expect(validScriptStub).to.have.been.calledOnce
@@ -1662,7 +676,7 @@ describe('./lib/lambda/handler.js', () => {
       expect(runPerformanceStub).to.not.have.been.called
     })
     it('calls performance mode if no mode is specified', () => {
-      script = validScript()
+      script = tagScript()
       handler.api.run(script, tagContext, noop)
       expect(validScriptStub).to.have.been.calledOnce
       expect(runAcceptanceStub).to.not.have.been.called
@@ -1670,7 +684,7 @@ describe('./lib/lambda/handler.js', () => {
     })
     it('detects performance mode declared as "perf"', () => {
       const mode = 'perf'
-      script = validScript()
+      script = tagScript()
       script.mode = mode
       handler.api.run(script, tagContext, noop)
       expect(validScriptStub).to.have.been.calledOnce
@@ -1679,7 +693,7 @@ describe('./lib/lambda/handler.js', () => {
     })
     it('detects performance mode declared as "performance"', () => {
       const mode = 'performance'
-      script = validScript()
+      script = tagScript()
       script.mode = mode
       handler.api.run(script, tagContext, noop)
       expect(validScriptStub).to.have.been.calledOnce
